requirements.txt
openpyxl>=3.0.0
pandas>=1.0.0
python-bidi>=0.4.2
